+++ plugins/node/opentelemetry-instrumentation-mysql2/src/utils.ts
@@ -131,7 +131,11 @@ export function getDbStatement(
 export function getSpanName(query: string | Query | QueryOptions): string {
   const rawQuery = typeof query === 'object' ? query.sql : query;
   // Extract the SQL verb
-  return rawQuery?.split(' ')?.[0];
+  const firstSpace = rawQuery?.indexOf(' ');
+  if (typeof firstSpace === 'number' && firstSpace !== -1) {
+    return rawQuery?.substring(0, firstSpace);
+  }
+  return rawQuery;
 }
 
 export const once = (fn: Function) => {