+++ plugins/node/instrumentation-socket.io/CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changelog
 
+## [0.43.0](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/instrumentation-socket.io-v0.42.0...instrumentation-socket.io-v0.43.0) (2024-10-25)
+
+
+### Features
+
+* update "@opentelemetry/*" dependencies to 1.27.0/0.54.0 ([2822511](https://github.com/open-telemetry/opentelemetry-js-contrib/commit/2822511a8acffb875ebd67ff2cf95980a9ddc01e))
+
+
+### Bug Fixes
+
+* change span names for socket-io ([#2495](https://github.com/open-telemetry/opentelemetry-js-contrib/issues/2495)) ([86dba74](https://github.com/open-telemetry/opentelemetry-js-contrib/commit/86dba74c5f12a3e47c08fef7cdb2edf0776ee6f8))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @opentelemetry/contrib-test-utils bumped from ^0.41.0 to ^0.42.0
+
 ## [0.42.0](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/instrumentation-socket.io-v0.41.0...instrumentation-socket.io-v0.42.0) (2024-09-02)