+++ plugins/node/opentelemetry-instrumentation-pg/CHANGELOG.md
@@ -1,5 +1,52 @@
 # Changelog
 
+## [0.47.0](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/instrumentation-pg-v0.46.0...instrumentation-pg-v0.47.0) (2024-10-25)
+
+
+### Features
+
+* **instrumentation-pg:** add error type to db duration metric ([#2476](https://github.com/open-telemetry/opentelemetry-js-contrib/issues/2476)) ([9564380](https://github.com/open-telemetry/opentelemetry-js-contrib/commit/9564380782164c8d8da6cac675d76d2fb4cdf20b))
+* update "@opentelemetry/*" dependencies to 1.27.0/0.54.0 ([2822511](https://github.com/open-telemetry/opentelemetry-js-contrib/commit/2822511a8acffb875ebd67ff2cf95980a9ddc01e))
+
+
+### Bug Fixes
+
+* **instrumentation-pg:** ensure db.client.operation.duration metric is recorded for Promises API usage of pg ([#2480](https://github.com/open-telemetry/opentelemetry-js-contrib/issues/2480)) ([97a2956](https://github.com/open-telemetry/opentelemetry-js-contrib/commit/97a2956f2703ebe62c16863decada93fef4574ec)), closes [#2380](https://github.com/open-telemetry/opentelemetry-js-contrib/issues/2380)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @opentelemetry/contrib-test-utils bumped from ^0.41.0 to ^0.42.0
+
+## [0.46.0](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/instrumentation-pg-v0.45.1...instrumentation-pg-v0.46.0) (2024-10-16)
+
+
+### Features
+
+* **instrumentation-pg:** add operation name attribute to db duration metric ([#2475](https://github.com/open-telemetry/opentelemetry-js-contrib/issues/2475)) ([b043ffb](https://github.com/open-telemetry/opentelemetry-js-contrib/commit/b043ffbe17947409c0ae9ea6fabe6ab1ec5b4cc9))
+
+
+### Bug Fixes
+
+* **instrumentation-pg:** not add duplicate listeners to pg pool ([#2484](https://github.com/open-telemetry/opentelemetry-js-contrib/issues/2484)) ([33c093d](https://github.com/open-telemetry/opentelemetry-js-contrib/commit/33c093daf77153ea3d0ab8d7925772efd37bdebe))
+
+## [0.45.1](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/instrumentation-pg-v0.45.0...instrumentation-pg-v0.45.1) (2024-10-10)
+
+
+### Bug Fixes
+
+* add missing @opentelemetry/core dependency ([#2473](https://github.com/open-telemetry/opentelemetry-js-contrib/issues/2473)) ([4d66431](https://github.com/open-telemetry/opentelemetry-js-contrib/commit/4d66431a2ad480720bf6c33663a419ec1404cd33))
+
+## [0.45.0](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/instrumentation-pg-v0.44.0...instrumentation-pg-v0.45.0) (2024-10-08)
+
+
+### Features
+
+* **instrumentation-pg:** implementation of metric operation duration ([#2380](https://github.com/open-telemetry/opentelemetry-js-contrib/issues/2380)) ([050fee0](https://github.com/open-telemetry/opentelemetry-js-contrib/commit/050fee0551ccf0fa593b84030ab73e2701372b4b))
+* **instrumentation-pg:** initial implementation of DB metrics ([#2349](https://github.com/open-telemetry/opentelemetry-js-contrib/issues/2349)) ([12adb43](https://github.com/open-telemetry/opentelemetry-js-contrib/commit/12adb4354f09ade438cd96340bdfd1f715b5fed3))
+
 ## [0.44.0](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/instrumentation-pg-v0.43.0...instrumentation-pg-v0.44.0) (2024-09-02)